refactor: relocate unit test
