+++ test/functional/adduser/logout.js
