+++ test/unit/search.spec.js
